Do all the things!
Fixed mapping data to race index by comparing keystone IDs.

Fixed the Future Artifacts tab so it actually works.

Added some annoying workarounds throughout to account for Headdress of the First Shaman being the name of both an Orc item and a Draenor Clans item. WoD is an AU fanfic, Blizz, we get it, thanks.

Added an (extensible) special case for Canopic Jar, which is interesting, but only for what it (maybe) contains.

Added Spanish translations.